refactor: remove all 'as any' type assertions for better type safety
- Replace 'as any' with proper types or 'as unknown as Type' where needed
- Use 'unknown' instead of 'any' for error parameters and logging functions
- Improve type safety throughout the codebase
- All tests passing with strict TypeScript checks
